// netrika/src/common/interfaces/user/UserRolesEnum.ts
export enum UserRolesEnum {
  RegistryAdmin = "RegistryAdminGroup",
  RegistryExpert = "RegistryExpertGroup",
  RegistrySuperExpert = "RegistrySuperExpertGroup",
  RegistryTherapist = "RegistryTherapistGroup",
  RegistryDiseaseKurator = "RegistryDiseaseKuratorGroup",
  RegistryOrgCurator = "RegistryOrgCuratorGroup",
  RegistrySuperUsr = "RegistrySuperUsrGroup",
}